Create basicUI
Implemented a basic layout structure using grid layout.

Currently, the program is able to switch images based on the combo box item selected. 
To incorporate the logic for switching camera feeds based on combo box, the program logic should be included in onComboSelected method.

Currently, the program just displays an alert box when reset button is clicked. 
The code for reset needs to be included in handleButton method.